Move test files to lib
