+++ .eslintrc.yml
@@ -1,4 +1,8 @@
-extends: standard
+extends:
+  - standard
+  - plugin:jest/recommended
+
+plugins: [ jest ]
 
 rules:
   arrow-parens: [2, as-needed]
@@ -12,3 +16,8 @@ rules:
   dot-location: [2, property]
   prefer-arrow-callback: 2
   prefer-const: 2
+
+overrides:
+  files: "*.test.js"
+  env:
+    jest/globals: true
