+++ lib/basic-auth.test.js
@@ -1,4 +1,4 @@
-const koala = require('../lib');
+const koala = require('.');
 const request = require('supertest');
 
 describe('Basic Auth', () => {
