+++ lib/headers.test.js
@@ -1,4 +1,4 @@
-const koala = require('../lib');
+const koala = require('.');
 const request = require('supertest');
 
 describe('Set headers', () => {
