--- test/.eslintrc.yml
@@ -1,6 +1,0 @@
-extends: [ plugin:jest/recommended ]
-
-env:
-  jest/globals: true
-
-plugins: [ jest ]
